pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "StudyBuddy - AI Study Assistant (Interview Code Exercise)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.1",
    "pydantic>=2.0.0",
    "typing_extensions>=4.0.0",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.4.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
markers = [
    "eval: marks tests as LLM evaluation tests (deselect with '-m \"not eval\"')",
]
# By default, skip eval tests (they require LLM API calls and are slower)
addopts = "-m 'not eval'"
testpaths = ["tests"]
pythonpath = ["."]
